refactor: upgrade express adapter
